+++ Widgets/Chart.php
@@ -14,6 +14,7 @@ use exface\Core\Exceptions\Widgets\WidgetConfigurationError;
 use exface\Core\Interfaces\Widgets\iFillEntireContainer;
 use exface\Core\Interfaces\Widgets\iLayoutWidgets;
 use exface\Core\CommonLogic\Model\Attribute;
+use exface\Core\CommonLogic\Traits\WidgetLayoutTrait;
 
 /**
  * A Button is the primary widget for triggering actions.
@@ -25,6 +26,8 @@ use exface\Core\CommonLogic\Model\Attribute;
  */
 class Chart extends AbstractWidget implements iShowDataSet, iHaveButtons, iHaveTopToolbar, iHaveBottomToolbar, iSupportLazyLoading, iFillEntireContainer, iLayoutWidgets
 {
+    
+    use WidgetLayoutTrait;
 
     /**
      *
@@ -79,12 +82,6 @@ class Chart extends AbstractWidget implements iShowDataSet, iHaveButtons, iHaveT
 
     /** @var string */
     private $lazy_loading_group_id = null;
-
-    private $number_of_columns = null;
-
-    private $column_stack_on_smartphones = null;
-
-    private $column_stack_on_tablets = null;
 
     const AXIS_X = 'x';
 
@@ -677,96 +674,6 @@ class Chart extends AbstractWidget implements iShowDataSet, iHaveButtons, iHaveT
     public function getAlternativeContainerForOrphanedSiblings()
     {
         return null;
-    }
-
-    /**
-     *
-     * {@inheritdoc}
-     *
-     * @see \exface\Core\Interfaces\Widgets\iLayoutWidgets::getNumberOfColumns()
-     */
-    public function getNumberOfColumns()
-    {
-        if (is_null($this->number_of_columns)) {
-            $widget = $this;
-            while ($widget->getParent()){
-                $widget = $widget->getParent();
-                if ($widget instanceof iLayoutWidgets && $widget->getNumberOfColumns()){
-                    $this->number_of_columns = $widget->getNumberOfColumns();
-                    break;
-                }
-            }
-            if (is_null($this->number_of_columns)) {
-                $this->number_of_columns = 4;
-            }
-    
-            $dimension = $this->getWidth();
-            if ($dimension->isRelative()) {
-                $width = $dimension->getValue();
-                if ($width === 'max') { $width = $this->number_of_columns; }
-                if ($width < 1) { $width = 1; }
-                if ($width < $this->number_of_columns) { $this->number_of_columns = $width; }
-            }
-        }
-        return $this->number_of_columns;
-    }
-
-    /**
-     *
-     * {@inheritdoc}
-     *
-     * @see \exface\Core\Interfaces\Widgets\iLayoutWidgets::setNumberOfColumns()
-     */
-    public function setNumberOfColumns($value)
-    {
-        $this->number_of_columns = intval($value);
-        return $this;
-    }
-
-    /**
-     *
-     * {@inheritdoc}
-     *
-     * @see \exface\Core\Interfaces\Widgets\iLayoutWidgets::getStackColumnsOnTabletsSmartphones()
-     */
-    public function getStackColumnsOnTabletsSmartphones()
-    {
-        return $this->column_stack_on_smartphones;
-    }
-
-    /**
-     *
-     * {@inheritdoc}
-     *
-     * @see \exface\Core\Interfaces\Widgets\iLayoutWidgets::setStackColumnsOnTabletsSmartphones()
-     */
-    public function setStackColumnsOnTabletsSmartphones($value)
-    {
-        $this->column_stack_on_smartphones = BooleanDataType::parse($value);
-        return $this;
-    }
-
-    /**
-     *
-     * {@inheritdoc}
-     *
-     * @see \exface\Core\Interfaces\Widgets\iLayoutWidgets::getStackColumnsOnTabletsTablets()
-     */
-    public function getStackColumnsOnTabletsTablets()
-    {
-        return $this->column_stack_on_tablets;
-    }
-
-    /**
-     *
-     * {@inheritdoc}
-     *
-     * @see \exface\Core\Interfaces\Widgets\iLayoutWidgets::setStackColumnsOnTabletsTablets()
-     */
-    public function setStackColumnsOnTabletsTablets($value)
-    {
-        $this->column_stack_on_tablets = BooleanDataType::parse($value);
-        return $this;
     }
 
     /**
